chore: clippy + doc audit pass
- math: suppress approx_constant/large_const_arrays on generated LUT;
  allow unnecessary_cast (needed in 64-bit Inner modes); allow
  suspicious_{op_assign,arithmetic}_impl (correct fixed-point scaling);
  allow should_implement_trait on shr (ergonomic inherent method kept)
- gameplay: add mut_from_ref allows on intrusive-list traversal methods;
  add repr(u8) to DoorKind/FloorKind/CeilKind; explicit transmute types
  in save.rs; drop redundant *mut T as *mut T no-op casts
- gamestate: drop redundant unsafe casts to load/save_game_from_bytes
- sound/rodio: replace index loops with position()+single-write pattern
- pic-data: collapse nested if-let chains; allow too_many_arguments on
  private build_columns helper; fix empty-line module doc comments
- rbsp: use &mut [T] instead of &mut Vec<T> in compact_vertices; allow
  unnecessary_cast (needed when Float=f32)
- wad/voxel: convert misplaced /// to //! module-level doc comments
- voxel-viewer: explicit lifetime on transmute::<Pixels<'_>, Pixels<'static>>

+++ tools/rbsp/src/polygon.rs
@@ -86,18 +86,18 @@ pub fn clip_convex_poly(
         );
 
         if crossing {
-            let cx = pool.vertices[curr_idx as usize].x as f64;
-            let cy = pool.vertices[curr_idx as usize].y as f64;
-            let nx = pool.vertices[next_idx as usize].x as f64;
-            let ny = pool.vertices[next_idx as usize].y as f64;
+            let cx = pool.vertices[curr_idx as usize].x;
+            let cy = pool.vertices[curr_idx as usize].y;
+            let nx = pool.vertices[next_idx as usize].x;
+            let ny = pool.vertices[next_idx as usize].y;
             let edx = nx - cx;
             let edy = ny - cy;
-            let pdx = partition.dx as f64;
-            let pdy = partition.dy as f64;
+            let pdx = partition.dx;
+            let pdy = partition.dy;
 
             let denom = pdx * edy - pdy * edx;
-            if denom.abs() > PARALLEL_EPSILON as f64 {
-                let num = pdy * (cx - px as f64) - pdx * (cy - py as f64);
+            if denom.abs() > PARALLEL_EPSILON {
+                let num = pdy * (cx - px) - pdx * (cy - py);
                 let u = (num / denom).clamp(0.0, 1.0);
                 let hx = (cx + u * edx) as Float;
                 let hy = (cy + u * edy) as Float;
